Add pnacl-freeze to the tests_lit/lit.cfg. Also, unsigned vs signed.
Otherwise, I don't have pnacl-freeze in my path, and
I think the lit tests have trouble finding it.

src/PNaClTranslator.cpp: In member function ‘uint32_t {anonymous}::FunctionParser::convertRelativeToAbsIndex(int32_t)’:
src/PNaClTranslator.cpp:882:55: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     if (Id > 0 && AbsNextId < static_cast<uint32_t>(Id)) {
                                                       ^
BUG=none
R=kschimpf@google.com, stichnot@chromium.org

Review URL: https://codereview.chromium.org/515003004

+++ src/PNaClTranslator.cpp
@@ -879,7 +879,7 @@ private:
   // an absolute value index.
   uint32_t convertRelativeToAbsIndex(int32_t Id) {
     int32_t AbsNextId = CachedNumGlobalValueIDs + LocalOperands.size();
-    if (Id > 0 && AbsNextId < static_cast<uint32_t>(Id)) {
+    if (Id > 0 && AbsNextId < Id) {
       std::string Buffer;
       raw_string_ostream StrBuf(Buffer);
       StrBuf << "Invalid relative value id: " << Id
